buscar usuario por id de int para Long, e alteracao no controller de conta para puxar objeto pelo id de usuario na requisicao http

+++ src/main/java/com/cesurg/projetopagamento/core/domain/usecase/UsuarioUseCaseImpl.java
@@ -55,5 +55,5 @@ public class UsuarioUseCaseImpl implements UsuarioUseCase {
     }
 
     @Override
-    public Usuario buscarPorId(int id) { return usuarioRepository.buscarPorId(id); }
+    public Usuario buscarPorId(Long id) { return usuarioRepository.buscarPorId(id); }
 }
